+++ internal/component/pyroscope/scrape/manager.go
@@ -8,6 +8,7 @@ import (
 	"github.com/go-kit/log"
 	"github.com/grafana/alloy/internal/alloy/logging/level"
 	"github.com/grafana/alloy/internal/component/pyroscope"
+	config_util "github.com/prometheus/common/config"
 	"github.com/prometheus/prometheus/discovery/targetgroup"
 )
 
@@ -15,6 +16,8 @@ var reloadInterval = 5 * time.Second
 
 type Manager struct {
 	logger log.Logger
+
+	httpClientOptions []config_util.HTTPClientOption
 
 	graceShut  chan struct{}
 	appendable pyroscope.Appendable
@@ -27,17 +30,18 @@ type Manager struct {
 	triggerReload chan struct{}
 }
 
-func NewManager(appendable pyroscope.Appendable, logger log.Logger) *Manager {
+func NewManager(o []config_util.HTTPClientOption, appendable pyroscope.Appendable, logger log.Logger) *Manager {
 	if logger == nil {
 		logger = log.NewNopLogger()
 	}
 	return &Manager{
-		logger:        logger,
-		appendable:    appendable,
-		graceShut:     make(chan struct{}),
-		triggerReload: make(chan struct{}, 1),
-		targetsGroups: make(map[string]*scrapePool),
-		targetSets:    make(map[string][]*targetgroup.Group),
+		httpClientOptions: o,
+		logger:            logger,
+		appendable:        appendable,
+		graceShut:         make(chan struct{}),
+		triggerReload:     make(chan struct{}, 1),
+		targetsGroups:     make(map[string]*scrapePool),
+		targetSets:        make(map[string][]*targetgroup.Group),
 	}
 }
 
@@ -88,7 +92,7 @@ func (m *Manager) reload() {
 	var wg sync.WaitGroup
 	for setName, groups := range m.targetSets {
 		if _, ok := m.targetsGroups[setName]; !ok {
-			sp, err := newScrapePool(m.config, m.appendable, log.With(m.logger, "scrape_pool", setName))
+			sp, err := newScrapePool(m.httpClientOptions, m.config, m.appendable, log.With(m.logger, "scrape_pool", setName))
 			if err != nil {
 				level.Error(m.logger).Log("msg", "error creating new scrape pool", "err", err, "scrape_pool", setName)
 				continue